feat(autoalign): add autoUpdate functionality
